% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ann_colors.R
\name{get_ann_colors}
\alias{get_ann_colors}
\title{Get color names from unique groups in data frame}
\usage{
get_ann_colors(df, ann_cols, pal = "")
}
\arguments{
\item{df}{A data frame with the columns of interest}

\item{ann_cols}{The columns of interest (get colors for)}

\item{pal}{A named list. Existing colors in palette}
}
\value{
A vector with color names from unique groups in data frame
}
\description{
Get color names from unique groups in data frame
}
