% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_gradient.R
\name{get_col_gradient}
\alias{get_col_gradient}
\title{Get color gradient}
\usage{
get_col_gradient(colRamp, n = 50)
}
\arguments{
\item{colRamp}{colorRampPalette functions object containing colors for palette. Use function get_col_palette() to create.}

\item{n}{numeric. Number of colors in gradient}
}
\value{
A character vector, names of colors (hex).
}
\description{
Get color gradient for a colorRampPalette with n number of colors.
}
\examples{
x <- LETTERS[1:6]
# The three lines of code below are equivalent
get_col_gradient(get_col_palette("RdBu"), 100)
get_col_palette("RdBu")(100)
colorRampPalette(RColorBrewer::brewer.pal(9, "RdBu"))(100)
}
