% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_palette.R
\name{get_col_palette}
\alias{get_col_palette}
\title{Get color palette in the form of colorRampPalette functions.}
\usage{
get_col_palette(brew_pal = NA, custom = NA, rev = F)
}
\arguments{
\item{brew_pal}{character string  or NA. Exact name of RColorBrewer palette. e.g. "RdBu".}

\item{custom}{Na or character vector with colors specifying palette order. e.g. c("red", "white", "blue"). Any hex codes or R colors (see colors()) can be used.}

\item{rev}{logical; should the brew_pal color palette order be reversed?}
}
\value{
A colorRampPalette color interpolation functions.
}
\description{
Get color palette in the form of colorRampPalette color interpolation functions. Use with get_col_gradient() or get_element_colors() from this package.
}
\details{
Only one of the parameters (brew_pal or custom) must be specified (not NA), otherwise function will use custom only.
}
\examples{
# Make a color palette of red to blue (RColorBrewer)
# The two lines of code below are equivalent
get_col_palette(brew_pal = "RdBu")
colorRampPalette(RColorBrewer::brewer.pal(9, "RdBu"))
#' # Reverse the palette (i.e. blue to red)
get_col_palette(brew_pal = "RdBu", rev = T)

# Make a custom color palette of forestgreen to darkorchid4 (RColorBrewer)
# The two lines of code below are equivalent
get_col_palette(custom = c("forestgreen", "darkorchid4"))
colorRampPalette(c("forestgreen", "darkorchid4"))
}
\seealso{
colorRampPalette function from {grDevices}. See ?colorRampPalette.
}
