% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_colors.R
\name{get_element_colors}
\alias{get_element_colors}
\title{Get colors for a named vector based on a color palette.}
\usage{
get_element_colors(v, colRamp, rearr = F)
}
\arguments{
\item{v}{character vector. Unique elements to get colors for.}

\item{colRamp}{colorRampPalette functions object containing colors for palette. Use function get_col_palette() to create. See ?colorRampPalette .}

\item{rearr}{logical; should vector be arranged so a color gradient isn't created with the original vector order?}
}
\value{
A character vector of colors with the same length as v. The names of this vector are the unique elements specified by v.
}
\description{
Get colors for a named vector based on a color ramp palette.
}
\examples{
x <- LETTERS[1:6]
get_element_colors(x, get_col_palette("RdBu"))
get_element_colors(x, colorRampPalette(RColorBrewer::brewer.pal(9, "RdBu")))
}
